requirements.txt
alembic==1.5.4
autopep8==1.5.5
bidict==0.21.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Migrate==2.6.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycodestyle==2.6.0
PyJWT==2.3.0
pymongo==3.11.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.3.0
python-socketio==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
